Langchain/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
chromadb
wikipedia
arxiv
langchainhub
groq
langchain_groq
beautifulsoup4
cassio
sentence_transformers